requirements.txt
Pillow==9.5.0
ultralytics==8.0.0
opencv-python==4.7.0.72
piexif==1.1.3
scikit-learn==1.2.2
torch
torchvision
numpy==1.24.3
matplotlib==3.7.1
